cloud/lsm: import laser_scan_matcher code from scan_tools
laser_scan_matcher node (the backing engine for kale_bot_lsm
node) has diverged so much with our own hacks that it
no longer makes sense to maintain it as a fork of the
upstream project (ROS/scan_tools).

This patch imports everything and only what we care
about from that package and prepares the ground for us
maintaining as part of the kale_bot project.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>