cloud/lsm: handle "leaking" when the pixel is too close
If the map pixel is too close to the scanner, one pixel will be
visible under a wide angle. If that aperture is wider than the
angular granularity of LIDAR the scan array will be populated
sparsely. This will furter cause the effect of beam reaching
through the solid surface and "bouncing" off some object that
is normally occluded.

We handle the problem by calculating the aperture angle under
which a pixel is visible and populating all scan slots
corresponding to the aperture with intensity and obseved range.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>