cloud/lsm: implement scan downsampling
When scan_downsample_rate parameter is set to a value greater
than 1, the scan frames are downsampled by the given rate.
Useful when the CPU is overloaded (and thus drops frames).
It is better in this case to drop in controlled manner.
Dropping of scan frames is done based on the frame sequence
number.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>