cloud/lsm: pass the scan timestamp to constructed scan
When constructing the scan, it is important that the
constructed scan time matches the time of the actual
scan that we are comparing to. This patch is the first
in the series intended to get this right. Pass the
timestamp of the real scan to the construct_scan_ function.
For now we only use it to stamp it back when publishing
the constructed scan.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>