cloud/lsm: transform the predicted pose covariance
Predicted pose is in map frame, but we do prediction relative
to the last output pose update cycle (essentially the
robot body frame as of last update). Covariance is therefore
determined in that frame, so we need to transform it into
the map frame in which we will do the fusion.

Signed-off-by: Ilija Hadzic <ilija.hadzic@nokia-bell-labs.com>