lsm: improve the accuracy of the scan construction algorithm
Instead of assuming all beams which hit a grid-cell have the length of
the distance to the center of the grid-cell, calculate each beam's
intersection point with the edge of the cell as though it were a solid
square.

This fixes a problem where close-together points would form jagged
patterns along straight lines, causing the correspondence normal vectors
to not actually align with the normal to the wall.

This change seems to greatly improve performance of LSM